requirements.txt
os
faiss-cpu
numpy
sentence_transformersr
mistralai